updated implementation
